v3
We added translator class to the project! It takes input array String and outputs to us translated Strings in the array on specific language! We got only German and English but in the future we can add others too!

We created a method to convert given input for an example "english" to "EN" because our DeepL API requires this specific input to translate the sentences!

Co-Authored-By: erminmiro <102598749+erminmiro@users.noreply.github.com>